split inclusion_dependency.py and check type for singular type in config
